splitted some code in main.rs
